RPC Authorization
This commit adds permission policy for raft consensus and adds
permission policy to config(but hides it from user).

It proposes minimum set of RPC methods which we have to allow peers in
order for all cluster features to work as expected. It restricts peers
from accessing methods that are not part of the proposed RPC method set.

Proposed authorization would work only for raft consensus.
`authorizeWithPolicy` would need to be edited in order to support
upcoming consensus mechanisms.

Issue #666